requirements.txt
amqp==5.1.1
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coverage==7.2.5
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.3
django-dbbackup==4.0.2
django-import-export==2.7.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
flake8==6.0.0
gunicorn==20.1.0
iniconfig==1.1.1
isort==5.12.0
kombu==5.2.4
MarkupPy==1.14
mccabe==0.7.0
mypy-extensions==1.0.0
odfpy==1.4.1
openpyxl==3.0.9
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pluggy==1.0.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.7
pytest==7.1.1
python-decouple==3.6
pytils==0.4.1
pytz==2022.1
PyYAML==6.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.2
tablib==3.2.0
tomli==2.0.1
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
